Livemap Command Fix
Fixed the /livemap command to correctly toggle player visibility on the
website livemaps; database updates correctly. I will reflect this in the
WebUI soon. I also fixed a null error that occurs when a player 'has not
set up their steam profile' resulting in a null value. I added a check
for this and a config option for the default (missing) avatar.